Create WriteFilesToSQL
This program opens a connection to the recipebook database in MYSQL, opens a directory
and reads the names and contents of the files within the directory. The program then
reformats these strings as appropriate in order to be parsed in a MySQL query.
The contents of the directory are then written to the database.